refactor: streamline useCounter and useTasks hooks by consolidating state management and removing unused handlers
