+++ src/hooks/useTasks.ts
@@ -1,4 +1,5 @@
 import { useState, useEffect } from "react";
+import { formatInputTime } from "../utils/formatsTimer";
 
 interface Task {
   task: string;
@@ -8,7 +9,25 @@ interface Task {
 
 export default function useTasks() {
   const [tasks, setTasks] = useState<Task[]>([]);
+  const [inputTask, setInputTask] = useState<string>("");
+  const [time, setTime] = useState<number>(0);
+  const [inputTime, setInputTime] = useState<string>("0:00:00");
 
+  const handleTimeChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const rawInput = event.target.value;
+    const formattedInput = formatInputTime(rawInput);
+    setInputTime(formattedInput);
+
+    const [hours, minutes, seconds] = formattedInput.split(":").map(Number);
+    const timeInSeconds = hours * 3600 + minutes * 60 + seconds;
+    setTime(timeInSeconds);
+  };
+
+  const handleTaskChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const task = event.target.value;
+    setInputTask(task);
+  }
+  
   const addTask = (task: string, time: string) => {
     if (task && time) {
       const newTask = { task, time, checked: false };
@@ -66,5 +85,13 @@ export default function useTasks() {
     addTask,
     handleDelete,
     handleCheckboxChange,
+    handleTimeChange,
+    handleTaskChange,
+    inputTask,
+    inputTime,
+    setInputTask,
+    setInputTime,
+    time,
+    setTime,
   };
 }
